Add files via upload
This is a mock website I have built. The idea is to keep track of how much profit or loss can be accumulated by buying a scratch card every day for a year.
I had videos I would like to add to this upload but GitHub cannot handle the size of them. So I have left them out.